* Removed the executable flag from setup.py to avoid accidentally    running it :)  * Excluded build/.svn from the list of autocleaned directories :)

+++ trunk/setup.py
@@ -64,6 +64,8 @@ except SystemExit, msg:
 
 # Cleanup
 for d in os.listdir ("build"):
+    if d [0] == ".":
+        continue
     fulld = os.path.join ("build", d)
     if os.path.isdir (fulld):
         RemoveDir (fulld)
